+++ src/modules/manufacturing/entities/kp-oem-mesin-sortir.ts
@@ -1,9 +1,9 @@
 import { Field, ObjectType } from 'type-graphql';
 import { BaseEntity, Column, Entity, PrimaryColumn } from 'typeorm';
 
-@Entity('VKY_KP_OEM_MESIN_SORTIR_V')
+@Entity('VKY_KP_OEM_MESIN_SORTIR')
 @ObjectType()
-export class KpOemMesinSortirView extends BaseEntity {
+export class KpOemMesinSortir extends BaseEntity {
   @Field()
   @PrimaryColumn({ name: 'CONTRACT' })
   contract!: string;
@@ -21,18 +21,14 @@ export class KpOemMesinSortirView extends BaseEntity {
   quantity!: number;
 
   @Field()
-  @Column({ name: 'QTY_OOREC' })
-  qtyOorec!: number;
-
-  @Field()
-  @Column({ name: 'QTY_SORTIR' })
-  qtySortir!: number;
-
-  @Field()
-  @Column({ name: 'QTY_OESHIP' })
-  qtyOeship!: number;
-
-  @Field()
   @Column({ name: 'STATUS' })
   status!: string;
+
+  @Field()
+  @Column({ name: 'CREATED_DATE' })
+  createdDate!: Date;
+
+  @Field()
+  @Column({ name: 'MODIFIED_DATE' })
+  modifiedDate!: Date;
 }