+++ src/modules/manufacturing/kp-oem-mesin-sortir.in.ts
@@ -1,9 +1,9 @@
 import { IsDate, IsNumber, MaxLength } from 'class-validator';
 import { Field, InputType } from 'type-graphql';
-import { KpOemSementara } from './entities/kp-oem-sementara';
+import { KpOemMesinSortir } from './entities/kp-oem-mesin-sortir';
 
 @InputType()
-export class KpOemSementaraInput implements Partial<KpOemSementara> {
+export class KpOemMesinSortirInput implements Partial<KpOemMesinSortir> {
   @Field()
   @MaxLength(5)
   contract!: string;
@@ -16,20 +16,12 @@ export class KpOemSementaraInput implements Partial<KpOemSementara> {
   @MaxLength(70)
   lotBatchNo!: string;
 
-  @Field({ nullable: true })
-  @MaxLength(50)
-  seriBeam?: string;
-
   @Field()
   @IsNumber()
-  pickAwal!: number;
+  quantity!: number;
 
   @Field()
-  @IsNumber()
-  pickAkhir!: number;
-
-  @Field()
-  @MaxLength(10)
+  @MaxLength(15)
   status!: string;
 
   @Field()
@@ -37,14 +29,6 @@ export class KpOemSementaraInput implements Partial<KpOemSementara> {
   createdDate!: Date;
 
   @Field()
-  @MaxLength(5)
-  createdAt!: string;
-
-  @Field()
   @IsDate()
   modifiedDate!: Date;
-
-  @Field()
-  @MaxLength(5)
-  modifiedAt!: string;
 }